Added utils codes to pipe_utils file.
